+++ src/main/java/com/immobilier/gestionImmobiliere/modules/biens/dto/requests/ModifierDureeLocationDTO.java
@@ -3,6 +3,7 @@ package com.immobilier.gestionImmobiliere.modules.biens.dto.requests;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -10,7 +11,7 @@ public class ModifierDureeLocationDTO {
     @NotNull private LocalDateTime nouvelleDateFin;
 
     // Renseigné uniquement si la prolongation génère un paiement complémentaire
-    private Double montantComplement;
+    private BigDecimal montantComplement;
     private String modePaiementComplement;
     private String referencePaiementComplement;
 }
