+++ src/main/java/com/immobilier/gestionImmobiliere/modules/biens/dto/requests/UpdateBienServiceDTO.java
@@ -2,12 +2,14 @@ package com.immobilier.gestionImmobiliere.modules.biens.dto.requests;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class UpdateBienServiceDTO {
     private Integer idSecteur;
     private Integer idCategorie;
     private String libelle;
     private String description;
-    private Double prixJournalier;
-    private Double prixMensuel;
+    private BigDecimal prixJournalier;
+    private BigDecimal prixMensuel;
 }
