+++ src/main/java/com/immobilier/gestionImmobiliere/modules/biens/dto/responses/PaiementLocationBienServiceResponseDTO.java
@@ -4,12 +4,13 @@ import com.immobilier.gestionImmobiliere.donnees.paiements.model.TypePaiementLoc
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data @Builder
 public class PaiementLocationBienServiceResponseDTO {
     private Integer idPaiement;
-    private Double montant;
+    private BigDecimal montant;
     private String modePaiement;
     private TypePaiementLocationBienService typePaiement;
     private LocalDateTime datePaiement;
