+++ src/main/java/com/immobilier/gestionImmobiliere/modules/contrats/dto/requests/CreateContratLocationDTO.java
@@ -3,6 +3,7 @@ package com.immobilier.gestionImmobiliere.modules.contrats.dto.requests;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -11,7 +12,7 @@ public class CreateContratLocationDTO {
     @NotNull private Integer idMaison;
     @NotNull private LocalDateTime dateEntree;
     private LocalDateTime dateSortie;
-    @NotNull private Double montantLoyer;
+    @NotNull private BigDecimal montantLoyer;
     private String typeContrat;
     private String etatDesLieuxEntree;
 }
