+++ src/main/java/com/immobilier/gestionImmobiliere/modules/paiements/dto/requests/CreatePaiementDTO.java
@@ -4,12 +4,13 @@ import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
 @Data
 public class CreatePaiementDTO {
-    @NotNull private Double montantPaiement;
+    @NotNull private BigDecimal montantPaiement;
     @NotNull private String modePaiement;
     private String referencePaiement;
     private LocalDateTime datePaiement;
