+++ src/main/java/com/immobilier/gestionImmobiliere/modules/paiements/dto/responses/PaiementResponseDTO.java
@@ -3,6 +3,7 @@ package com.immobilier.gestionImmobiliere.modules.paiements.dto.responses;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -10,7 +11,7 @@ import java.util.List;
 public class PaiementResponseDTO {
     private Integer idPaiement;
     private LocalDateTime datePaiement;
-    private Double montantPaiement;
+    private BigDecimal montantPaiement;
     private String sens;
     private String modePaiement;
     private String referencePaiement;
